requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gpt4all
sentence-transformers==2.2.2
torch==2.2.0
chromadb==0.4.18
pypdf2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi-cors==0.0.6
numpy   
openai
email-validator